Final Push - Ready for Review
Added in a video from youtube. Added in support for the video and audio files in case of the browser not supporting it. Added in two audio files, with progressive enhancement. Changes in CSS so that the audio has a background and is offset to the right with a shadow.